requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
psycopg2-binary>=2.9.0
qdrant-client>=1.7.0
pydantic>=2.0.0
python-multipart
requests
numpy
scikit-learn

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
pytest-mock>=3.10.0
httpx>=0.24.0
coverage>=7.0.0

# Development dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
